Refactor ast parser
