+++ app/electron/main/code/ast/templateNode.ts
@@ -1,6 +1,6 @@
 import traverse from '@babel/traverse';
 import t from '@babel/types';
-import { readCodeBlock } from '.';
+import { readCodeBlock } from '..';
 import { parseJsx } from './babel';
 import { TemplateNode, TemplateTag } from '/common/models/element/templateNode';
