+++ app/src/lib/editor/engine/insert/index.ts
@@ -109,7 +109,7 @@ export class InsertManager {
         const actionElement: ActionElement = {
             tagName: 'div',
             attributes: {
-                id: nanoid(),
+                [EditorAttributes.DATA_ONLOOK_UNIQUE_ID]: nanoid(),
                 [EditorAttributes.DATA_ONLOOK_INSERTED]: 'true',
                 [EditorAttributes.DATA_ONLOOK_TIMESTAMP]: Date.now().toString(),
             },
